+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ---
 ## [X.X.X]
-* Feature:
+* Features:
 	* Add plugin methods to reflet all player's states
+	* Remove useless MediaPlayer delegate attribute
 * Fix:
 	* Reset item duration when loading media